Changed crit chance/multiplier to float for weapons Changed RecordTimeElapsedOption to record ticksElapsed Fixed star piece bonus not showing in weapon when applied

+++ src/main/java/com/ebicep/warlords/abilties/Inferno.java
@@ -74,7 +74,7 @@ public class Inferno extends AbstractAbility {
             }
 
             @Override
-            public int addCritChanceFromAttacker(WarlordsDamageHealingEvent event, int currentCritChance) {
+            public float addCritChanceFromAttacker(WarlordsDamageHealingEvent event, float currentCritChance) {
                 if (event.getAbility().isEmpty() || event.getAbility().equals("Time Warp")) {
                     return currentCritChance;
                 }
@@ -83,7 +83,7 @@ public class Inferno extends AbstractAbility {
             }
 
             @Override
-            public int addCritMultiplierFromAttacker(WarlordsDamageHealingEvent event, int currentCritMultiplier) {
+            public float addCritMultiplierFromAttacker(WarlordsDamageHealingEvent event, float currentCritMultiplier) {
                 if (event.getAbility().isEmpty() || event.getAbility().equals("Time Warp")) {
                     return currentCritMultiplier;
                 }
