+++ src/main/java/com/ebicep/warlords/abilties/TimeWarp.java
@@ -119,7 +119,7 @@ public class TimeWarp extends AbstractAbility {
                 }
 
                 @Override
-                public int addCritChanceFromAttacker(WarlordsDamageHealingEvent event, int currentCritChance) {
+                public float addCritChanceFromAttacker(WarlordsDamageHealingEvent event, float currentCritChance) {
                     if (event.getAbility().isEmpty() || event.getAbility().equals("Time Warp")) {
                         return currentCritChance;
                     }
@@ -127,7 +127,7 @@ public class TimeWarp extends AbstractAbility {
                 }
 
                 @Override
-                public int addCritMultiplierFromAttacker(WarlordsDamageHealingEvent event, int currentCritMultiplier) {
+                public float addCritMultiplierFromAttacker(WarlordsDamageHealingEvent event, float currentCritMultiplier) {
                     if (event.getAbility().isEmpty() || event.getAbility().equals("Time Warp")) {
                         return currentCritMultiplier;
                     }
